core: fixed u32 type

+++ src/types.hpp
@@ -6,7 +6,7 @@
 #include <ranges>
 
 using i32 = std::int32_t;
-using u32 = std::uint64_t;
+using u32 = std::uint32_t;
 using u64 = std::uint64_t;
 
 namespace types {
